third_party/WebKit/LayoutTests/http/tests/storage/async_cookies/document_get_set.html

<!doctype html>
<meta charset="utf-8">
<title>Async Cookes: cookieStore.get() sees cookieStore.set() cookie</title>
<link rel="author" href="pwnall@chromium.org" title="Victor Costan">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
'use strict';

promise_test(async testCase => {
  await cookieStore.set('cookie-name', 'cookie-value');
  const cookie = await cookieStore.get('cookie-name');

  assert_equals(cookie.name, 'cookie-name');
  assert_equals(cookie.value, 'cookie-value');

  try { await cookieStore.delete('cookie-name'); } catch (e) {}
}, 'cookieStore.get returns the cookie written by cookieStore.set');

</script>
